refactor(logic): move work-phase handlers to dedicated files
Make runWorkPhase a thin dispatcher by moving handler interface/context and concrete target handlers into work_handlers/*, keeping behavior intact while improving per-handler testability for issue #1958 phase 2.

Made-with: Cursor
