+++ packages/colonizethis_logic/lib/src/orders/work_handlers/explore_work_handler.dart
@@ -1,7 +1,9 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/package_logger.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 import '../../constants.dart';
+import 'work_order_handler.dart';
 
 final _log = packageLogger();
 
@@ -46,4 +48,34 @@ bool tryApplyExploreWorkOrder({
     ),
   );
   return true;
+}
+
+class ExploreWorkOrderHandler implements WorkOrderHandler {
+  const ExploreWorkOrderHandler();
+
+  @override
+  bool supports(String target) => target == kWorkTargetExplore;
+
+  @override
+  bool tryApply(
+    WorkOrderExecutionContext context,
+    WorkOrder order,
+    Unit unit,
+    String targetTileKey,
+    bool hasValidTarget,
+  ) {
+    if (!isExplorerUnit(unit.type) ||
+        unit.currentWork != null ||
+        !hasValidTarget) {
+      return false;
+    }
+    return tryApplyExploreWorkOrder(
+      game: context.state.game,
+      order: order,
+      unit: unit,
+      targetTileKey: targetTileKey,
+      regionForUnit: context.regionForUnit,
+      updateUnit: context.updateUnit,
+    );
+  }
 }
